+++ src/main/java/me/lokka30/phantomworlds/commands/phantomworlds/subcommands/DebugSubcommand.java
@@ -3,7 +3,7 @@ package me.lokka30.phantomworlds.commands.phantomworlds.subcommands;
 import me.lokka30.microlib.messaging.MessageUtils;
 import me.lokka30.microlib.messaging.MultiMessage;
 import me.lokka30.phantomworlds.PhantomWorlds;
-import me.lokka30.phantomworlds.commands.ISubcommand;
+import me.lokka30.phantomworlds.commands.Subcommand;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.jetbrains.annotations.NotNull;
@@ -17,7 +17,7 @@ import java.util.Locale;
  * @author lokka30
  * @since v2.0.0
  */
-public class DebugSubcommand implements ISubcommand {
+public class DebugSubcommand implements Subcommand {
 
     /*
     cmd: /pw debug ...