+++ src/main/java/me/lokka30/phantomworlds/commands/phantomworlds/subcommands/InfoSubcommand.java
@@ -2,7 +2,7 @@ package me.lokka30.phantomworlds.commands.phantomworlds.subcommands;
 
 import me.lokka30.microlib.messaging.MultiMessage;
 import me.lokka30.phantomworlds.PhantomWorlds;
-import me.lokka30.phantomworlds.commands.ISubcommand;
+import me.lokka30.phantomworlds.commands.Subcommand;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.jetbrains.annotations.NotNull;
@@ -15,7 +15,7 @@ import java.util.List;
  * @author lokka30
  * @since v2.0.0
  */
-public class InfoSubcommand implements ISubcommand {
+public class InfoSubcommand implements Subcommand {
 
     /*
     cmd: /pw info