+++ src/main/java/me/lokka30/phantomworlds/commands/phantomworlds/subcommands/TeleportSubcommand.java
@@ -2,7 +2,7 @@ package me.lokka30.phantomworlds.commands.phantomworlds.subcommands;
 
 import me.lokka30.microlib.messaging.MultiMessage;
 import me.lokka30.phantomworlds.PhantomWorlds;
-import me.lokka30.phantomworlds.commands.ISubcommand;
+import me.lokka30.phantomworlds.commands.Subcommand;
 import me.lokka30.phantomworlds.misc.Utils;
 import org.bukkit.Bukkit;
 import org.bukkit.command.Command;
@@ -17,7 +17,7 @@ import java.util.*;
  * @author lokka30, madison-allen
  * @since v2.0.0
  */
-public class TeleportSubcommand implements ISubcommand {
+public class TeleportSubcommand implements Subcommand {
 
     /*
     cmd: /pw teleport <world> [player]